Rename GetPositions to GetAccountPositions.
