+++ Primary.Tests/AccountsTests.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using Primary.Data.Orders;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -37,6 +38,48 @@ namespace Primary.Tests
             Assert.That(account.Id, Is.Not.EqualTo(default));
             Assert.That(account.Name, Is.EqualTo(Api.DemoAccount));
             Assert.That(account.Status, Is.True);
+        }
+
+        [Test]
+        [Timeout(10000)]
+        public async Task PositionsCanBeRetrieved()
+        {
+            var marketData = await GetSomeMarketData();
+            TestContext.Out.WriteLine("GetSomeMarketData OK");
+            var symbol = marketData.InstrumentId.Symbol;
+
+            // Take the opposite side.
+            var order = new Order()
+            {
+                InstrumentId = marketData.InstrumentId,
+                Type = Type.Market,
+                Side = marketData.Data.Offers?.Length > 0 ? Side.Buy : Side.Sell,
+                Quantity = AllInstrumentsBySymbol[symbol].MinimumTradeVolume,
+            };
+
+            var orderId = await Api.SubmitOrder(Api.DemoAccount, order);
+            TestContext.Out.WriteLine("Api.SubmitOrder OK");
+            await WaitForOrderToComplete(orderId);
+            TestContext.Out.WriteLine("WaitForOrderToComplete OK");
+
+            var positions = await Api.GetAccountPositions(Api.DemoAccount);
+            TestContext.Out.WriteLine("Api.GetAccountPositions OK");
+            Assert.That(positions, Is.Not.Null);
+
+            var position = positions.FirstOrDefault(p => p.Symbol == symbol);
+            Assert.That(position, Is.Not.EqualTo(default));
+            Assert.That(position.Symbol, Is.EqualTo(symbol));
+
+            if (order.Side == Side.Buy)
+            {
+                Assert.That(position.BuySize, Is.GreaterThanOrEqualTo(order.Quantity));
+                Assert.That(position.OriginalBuyPrice, Is.Not.EqualTo(0));
+            }
+            else
+            {
+                Assert.That(position.SellSize, Is.GreaterThanOrEqualTo(order.Quantity));
+                Assert.That(position.OriginalSellPrice, Is.Not.EqualTo(0));
+            }
         }
     }
 }
